% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{crosswalk_sabs}
\alias{crosswalk_sabs}
\title{Crosswalk between 2 Sabs}
\usage{
crosswalk_sabs(
  conn,
  conn_fun = "pg13::local_connect()",
  sab1,
  sab2,
  to_schema,
  to_table_name,
  crosswalk_schema = "umls_crosswalk",
  verbose = TRUE,
  render_sql = TRUE
)
}
\arguments{
\item{to_schema}{(Optional). If provided along with a \code{to_table_name}, will
write the crosswalk to Postgres. This is useful for larger crosswalks that
would require too much memory in the R environment.}

\item{to_table_name}{(Optional).}
}
\description{
Write a crosswalk to a table or return into
the R environment as a dataframe.
}
\seealso{
\code{\link[stringr]{str_replace}}
\code{\link[glue]{glue}}
\code{\link[pg13]{send}},\code{\link[pg13]{c("query", "query")}}
}
