% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDDL.R
\name{deriveDDLFromDataElements}
\alias{deriveDDLFromDataElements}
\title{Derive DDL from Data Elements}
\usage{
deriveDDLFromDataElements(dataElementsDf)
}
\arguments{
\item{dataElementsDf}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
This function takes the data frame scraped by \code{getDataElements} and adds additional columns to generate new ddl scripts for UMLS Metathesaurus Tables.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{rownames}}
\code{\link[tidyr]{separate_rows}}
\code{\link[dplyr]{filter}},\code{\link[dplyr]{rename}},\code{\link[dplyr]{mutate}}
\code{\link[rubix]{call_mr_clean}}
\code{\link[stringr]{str_replace}}
}
