% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeSQL.R
\name{executeSQL}
\alias{executeSQL}
\title{Execute SQL}
\usage{
executeSQL(sqlPath, ..., conn)
}
\arguments{
\item{sqlPath}{path to source SQL file}

\item{...}{Parameter values, if any, that are passed to the SqlRender::render function.}
}
\description{
This function:
1. Renders a SQL File
2. Translates the dialect to Oracle dialect
3. To identify exact statements that have thrown errors and they are first isolated by splitting single string read from the file on the semicolon
4. Each single line is sent to the database. If an error is thrown, the single statement is printed in the R console in red.
}
