% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeSQL2.R
\name{executeSQL2}
\alias{executeSQL2}
\title{Execute SQL}
\usage{
executeSQL2(sqlPaths, ..., conn)
}
\arguments{
\item{...}{Parameter values, if any, that are passed to the SqlRender::render function.}

\item{sqlPath}{path to source SQL file}
}
\description{
This function:
1. Renders a SQL File
2. Translates the dialect to Oracle dialect
3. To identify exact statements that have thrown errors and they are first isolated by splitting single string read from the file on the semicolon
4. Each single line is sent to the database. If an error is thrown, the single statement is printed in the R console in red.
}
