% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxnorm_map.R
\name{get_rxnorm_paths}
\alias{get_rxnorm_paths}
\title{Retrieve the Table Listing RxNorm Paths}
\usage{
get_rxnorm_paths(
  url = "https://lhncbc.nlm.nih.gov/RxNav/applications/RxNavViews.html#label:appendix"
)
}
\arguments{
\item{url}{Source URL. It is not hardcoded in case the URL changes in the future.
Default: 'https://lhncbc.nlm.nih.gov/RxNav/applications/RxNavViews.html#label:appendix'}
}
\value{
Tibble of a machine-readable version of what is read at the \code{url} parameter.
}
\description{
Retrieve the table found at
\url{https://lhncbc.nlm.nih.gov/RxNav/applications/RxNavViews.html#label:appendix}
that will be used to generate the SQL queries to derive RxNorm maps.
}
\details{
The tibble parsed from the URL is cached indefinitely. This is to avoid losing
this dataframe in the event that the source documentation is ever lost.
}
\seealso{
Other RxNorm Map: 
\code{\link{get_rxnorm_ingredient_map}()},
\code{\link{get_rxnorm_map}()},
\code{\link{write_rxnorm_path_lookup}()}
}
\concept{RxNorm Map}
