% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxnorm_map.R
\name{get_rxnorm_tty_lookup}
\alias{get_rxnorm_tty_lookup}
\title{Get TTY Lookup}
\usage{
get_rxnorm_tty_lookup()
}
\description{
Get custom expansions from tty derived from the
MRDOC table as a data frame.
}
\seealso{
Other RxNorm Map: 
\code{\link{get_rxnorm_ingredient_map}()},
\code{\link{get_rxnorm_map}()},
\code{\link{read_rxnorm_paths}()},
\code{\link{setup_rxmap}()},
\code{\link{write_rxnorm_map}()},
\code{\link{write_rxnorm_path_lookup}()}
}
\concept{RxNorm Map}
