% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDDL.R
\name{getDataElements}
\alias{getDataElements}
\title{Get Data Elements}
\usage{
getDataElements()
}
\description{
Get the table found at \href{https://www.nlm.nih.gov/research/umls/knowledge_sources/metathesaurus/release/columns_data_elements.html} as a data frame
}
\seealso{
\code{\link[xml2]{read_xml}}
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_table}}
}
