% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMeta.R
\name{loadMeta}
\alias{loadMeta}
\title{Load META Tables}
\usage{
loadMeta(path, dbname = "umls", username, password, full = FALSE)
}
\arguments{
\item{path}{Path to the unpacked RRF files}
}
\description{
Load META Tables
}
