% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamorphoSys.R
\name{metamorphoSys}
\alias{metamorphoSys}
\title{Open Metamorphosys}
\usage{
metamorphoSys(path_to_mmsys)
}
\arguments{
\item{path_to_mmsys}{Path to \code{run_mac.sh} in unpacked Metamorphysis download}
}
\description{
Open Metamorphosys
}
