% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_data_defintion_functions.R
\name{mth_field_definitions}
\alias{mth_field_definitions}
\title{Get the definitions of al the Metathesaurus Fields}
\usage{
mth_field_definitions()
}
\description{
Sourced from \href{https://www.nlm.nih.gov/research/umls/knowledge_sources/metathesaurus/release/columns_data_elements.html}
}
\seealso{
\code{\link[tibble]{tribble}}
}
