% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql_ddl.R
\name{mysql_ddl}
\alias{mysql_ddl}
\title{DDL META Tables}
\usage{
mysql_ddl(
  dbname = "umls",
  username,
  password,
  mrconso_only = FALSE,
  omop_only = FALSE,
  english_only = TRUE
)
}
\arguments{
\item{path}{Path to the unpacked RRF files}
}
\description{
DDL META Tables
}
