% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openMetamorphysis.R
\name{openMetamorphysis}
\alias{openMetamorphysis}
\title{Open Metamorphysis}
\usage{
openMetamorphysis(path_to_mmsys)
}
\arguments{
\item{path_to_mmsys}{Path to \code{run_mac.sh} in unpacked Metamorphysis download}
}
\description{
Open Metamorphysis
}
