% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_ddl.R
\name{pg_ddl}
\alias{pg_ddl}
\title{DDL META Tables}
\usage{
pg_ddl(
  conn,
  mrconso_only = FALSE,
  omop_only = FALSE,
  english_only = TRUE,
  download_path
)
}
\arguments{
\item{path}{Path to the unpacked RRF files}
}
\description{
DDL META Tables
}
