% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populateMeta.R
\name{populateMeta}
\alias{populateMeta}
\title{Populate META Tables}
\usage{
populateMeta(path)
}
\arguments{
\item{path}{Path to the META subdirectory of unpacked UMLS Metathesaurus}
}
\description{
Populate META Tables
}
