% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_data_defintion_functions.R
\name{readable_mth_tables}
\alias{readable_mth_tables}
\title{List the Human Readable Names of MTH Tables}
\usage{
readable_mth_tables()
}
\description{
List the Human Readable Names of MTH Tables
}
\seealso{
\code{\link[tibble]{tribble}}
}
