% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mysql.R
\name{run_mysql}
\alias{run_mysql}
\title{Execute SQL}
\usage{
run_mysql(sqls, ..., conn)
}
\arguments{
\item{...}{Parameter values, if any, that are passed to the SqlRender::render function.}

\item{sqlPath}{path to source SQL file}
}
\description{
Execute SQL
}
