% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pg.R
\name{run_pg}
\alias{run_pg}
\title{Instantiate Postgres}
\usage{
run_pg(
  conn,
  schema = "mth",
  rrf_dir,
  mrconso_only = FALSE,
  omop_only = FALSE,
  english_only = TRUE
)
}
\arguments{
\item{conn}{Database connection. For MySQL5.5, the connection made by RMySQL's dbConnect and for Postgres, the connection via DatabaseConnector.}

\item{schema}{For Postgres executions only, the schema to write all the tables to, Default: 'mth'}

\item{rrf_dir}{Path to directory to the RRF Files either unpackaged by direct download or produced by Metamorphosys}

\item{mrconso_only}{Setup only the MRCONSO Table? Default: FALSE}

\item{omop_only}{Setup the MRCONSO, MRHIER, MRMAP, MRSMAP, MRSAT, AND MRREL Tables only? Default: FALSE}

\item{english_only}{Setup only the ENG Tables? PARAM_DESCRIPTION, Default: TRUE}
}
\description{
These functions setup the instance of the given DBMS. All SQL Statements that fail to execute are printed back in the console at the conclusion of execution. It is important to note that the schemas are dropped and rewritten with the run and all present tables will be lost.
}
\seealso{
\code{\link[pg13]{lsSchema}},\code{\link[pg13]{send}},\code{\link[pg13]{lsTables}},\code{\link[pg13]{dropTable}}
\code{\link[SqlRender]{render}}
\code{\link[tibble]{as_tibble}}
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}},\code{\link[dplyr]{distinct}}
\code{\link[stringr]{str_remove}}
\code{\link[progress]{progress_bar}}

Other setup: 
\code{\link{run_mysql}()}
}
\concept{setup}
