% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_setup.R
\name{run_postprocessing}
\alias{run_postprocessing}
\title{Run Postprocessing Only}
\usage{
run_postprocessing(
  conn,
  conn_fun = "pg13::local_connect()",
  mth_version,
  mth_release_dt,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\description{
Run Postprocessing Only
}
