% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_setup.R
\name{run_setup}
\alias{run_setup}
\title{Instantiate Postgres}
\usage{
run_setup(
  conn,
  conn_fun = "pg13::local_connect()",
  schema = "umls",
  path_to_rrfs,
  steps = c("reset_schema", "ddl_tables", "copy_rrfs", "add_indexes", "log",
    "setup_crosswalk"),
  postprocess = TRUE,
  mrconso_only = FALSE,
  omop_only = FALSE,
  english_only = TRUE,
  log_schema = "public",
  log_table_name = "setup_umls_log",
  log_version,
  log_release_date,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\arguments{
\item{conn}{Database connection via DatabaseConnector.}

\item{conn_fun}{Instead of supplying of a direct
database connection, an expression can be supplied that
automatically connects and disconnects the connection
once the function is run.}

\item{schema}{For Postgres executions only, the
schema to write all the tables to, Default: 'umls'}

\item{mrconso_only}{Setup only the MRCONSO Table? Default: FALSE}

\item{omop_only}{Setup the MRCONSO, MRHIER, MRMAP,
MRSMAP, MRSAT, AND MRREL Tables only? Default: FALSE}

\item{english_only}{Setup only the ENG Tables? Default: TRUE}
}
\description{
Instantiate Postgres
}
