% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIndices.R
\name{runIndices}
\alias{runIndices}
\title{Run Indices}
\usage{
runIndices(dbname = "umls", username, password)
}
\arguments{
\item{path}{Path to the unpacked RRF files}
}
\description{
Run indices on the loaded data.
}
