% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{setup_crosswalk_schema}
\alias{setup_crosswalk_schema}
\title{Write More Than 1 SAB Subset at Once}
\usage{
setup_crosswalk_schema(
  sabs,
  conn_fun = "pg13::local_connect()",
  crosswalk_schema = "mth_crosswalk"
)
}
\arguments{
\item{sabs}{(optional) Vector of SABs to write. If missing,
all the sabs in the existing MRCONSO table are written.}

\item{crosswalk_schema}{Schema to write these extension tables to. Default: 'mth_crosswalk'}
}
\value{
One or more tables of the sab subsets to the given schema. A new 'label'
field is added to designate the core concept name for the given
concept.
}
\description{
Write all or some of the SAB subset tables via
\code{write_crosswalk_table()}.
}
\seealso{
\code{\link[pg13]{drop_cascade}},\code{\link[pg13]{create_schema}}
}
