% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_extension.R
\name{setup_mrconso_sab_subsets}
\alias{setup_mrconso_sab_subsets}
\title{Write More Than 1 SAB Subset at Once}
\usage{
setup_mrconso_sab_subsets(
  sabs,
  conn_fun = "pg13::local_connect()",
  extension_schema = "mrconso_sab"
)
}
\arguments{
\item{sabs}{(optional) Vector of SABs to write. If missing,
all the sabs in the existing MRCONSO table are written.}

\item{extension_schema}{Schema to write these extension tables to. Default: 'mrconso_sab'}
}
\value{
One or more tables of the sab subsets to the given schema. A new 'label'
field is added to designate the core concept name for the given
concept.
}
\description{
Write all or some of the SAB subset tables via
\code{write_sab_subset_table()}.
}
\seealso{
\code{\link[pg13]{drop_cascade}},\code{\link[pg13]{create_schema}}
}
