% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{setup_mysql_mth}
\alias{setup_mysql_mth}
\title{Instantiate MySQL5.5}
\usage{
setup_mysql_mth(
  conn,
  mrconso_only = FALSE,
  omop_only = FALSE,
  english_only = TRUE,
  rrf_dir
)
}
\description{
Instantiate MySQL5.5
}
\seealso{
\code{\link[preQL]{query}}
\code{\link[purrr]{map}},\code{\link[purrr]{map2}}
\code{\link[SqlRender]{render}}
\code{\link[stringr]{str_replace}},\code{\link[stringr]{str_remove}}
\code{\link[RMySQL]{character(0)}}
\code{\link[tibble]{as_tibble}}
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}},\code{\link[dplyr]{distinct}}
\code{\link[rubix]{filter_for}}
\code{\link[progress]{progress_bar}}

Other setup: 
\code{\link{setup_pg_mth}()}
}
\concept{setup}
