% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{setup_mysql_umls}
\alias{setup_mysql_umls}
\title{Instantiate MySQL5.5}
\usage{
setup_mysql_umls(
  conn,
  mrconso_only = FALSE,
  omop_only = FALSE,
  english_only = TRUE,
  rrf_dir
)
}
\arguments{
\item{conn}{Database connection. For MySQL5.5,
the connection made by RMySQL's dbConnect and for Postgres,
the connection via DatabaseConnector.}

\item{mrconso_only}{Setup only the MRCONSO Table? Default: FALSE}

\item{omop_only}{Setup the MRCONSO, MRHIER, MRMAP,
MRSMAP, MRSAT, AND MRREL Tables only? Default: FALSE}

\item{english_only}{Setup only the ENG Tables? Default: TRUE}
}
\description{
Instantiate MySQL5.5
}
\seealso{
\code{\link[preQL]{query}}
\code{\link[purrr]{map}},\code{\link[purrr]{map2}}
\code{\link[SqlRender]{render}}
\code{\link[stringr]{str_replace}},\code{\link[stringr]{str_remove}}
\code{\link[RMySQL]{character(0)}}
\code{\link[tibble]{as_tibble}}
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}},\code{\link[dplyr]{distinct}}
\code{\link[rubix]{filter_for}}
\code{\link[progress]{progress_bar}}

Other setup: 
\code{\link{run_setup}()}
}
\concept{setup}
