% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg.R
\name{setup_pg_mth}
\alias{setup_pg_mth}
\title{Instantiate Postgres}
\usage{
setup_pg_mth(
  conn,
  schema = "mth",
  rrf_dir,
  mrconso_only = FALSE,
  omop_only = FALSE,
  english_only = TRUE
)
}
\description{
Instantiate Postgres
}
\seealso{
\code{\link[pg13]{lsSchema}},\code{\link[pg13]{send}},\code{\link[pg13]{lsTables}},\code{\link[pg13]{dropTable}}
\code{\link[SqlRender]{render}}
\code{\link[tibble]{as_tibble}}
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}},\code{\link[dplyr]{distinct}}
\code{\link[stringr]{str_remove}}
\code{\link[progress]{progress_bar}}

Other setup: 
\code{\link{setup_mysql_mth}()}
}
\concept{setup}
