% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxnorm_map.R
\name{setup_rxmap}
\alias{setup_rxmap}
\title{Setup RxMap}
\usage{
setup_rxmap(
  conn,
  conn_fun = "pg13::local_connect()",
  schema = "umls",
  target_schema = "rxmap",
  rxclass_schema = "umls_rxclass",
  umls_version,
  umls_release_dt,
  url = "https://lhncbc.nlm.nih.gov/RxNav/applications/RxNavViews.html#label:appendix",
  check_for_updates = FALSE,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  checks = ""
)
}
\description{
Setup RxMap
}
\seealso{
Other RxNorm Map: 
\code{\link{get_rxnorm_ingredient_map}()},
\code{\link{get_rxnorm_map}()},
\code{\link{get_rxnorm_tty_lookup}()},
\code{\link{read_rxnorm_paths}()},
\code{\link{write_rxnorm_map}()},
\code{\link{write_rxnorm_path_lookup}()}
}
\concept{RxNorm Map}
