% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tty}
\alias{tty}
\title{TTY Description}
\format{
A data frame with 241 rows and 2 variables:
\describe{
\item{\code{TTY}}{character TTY value}
\item{\code{TTY_Description}}{character TTY description}
}
}
\usage{
tty
}
\description{
TTY description from \href{https://www.nlm.nih.gov/research/umls/knowledge_sources/metathesaurus/release/abbreviations.html}{Metathesaurus Abbreviations Page}
}
\keyword{datasets}
