% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tty_class}
\alias{tty_class}
\title{TTY Classes}
\format{
A data frame with 289 rows and 2 variables:
\describe{
\item{\code{TTY_Class}}{character TTY Class}
\item{\code{TTY_Description}}{character TTY Description}
}
}
\usage{
tty_class
}
\description{
TTY classes described in href{https://www.nlm.nih.gov/research/umls/knowledge_sources/metathesaurus/release/abbreviations.html}{Metathesaurus Abbreviations Page}
}
\keyword{datasets}
