% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{write_crosswalk_table}
\alias{write_crosswalk_table}
\title{Write a Crosswalk Table}
\usage{
write_crosswalk_table(
  sab,
  crosswalk_schema = "umls_crosswalk",
  conn_fun = "pg13::local_connect(verbose = FALSE)",
  tty_ranking,
  log_schema = "public",
  log_table_name = "setup_umls_log",
  crosswalk_log_table_name = "setup_umls_crosswalk_log",
  verbose = TRUE,
  render_sql = TRUE
)
}
\arguments{
\item{sab}{Vocabulary.}

\item{crosswalk_schema}{Schema to which these extension tables are to be written.}

\item{tty_rank}{(Optional). Vector of case insensitive tty values for the
given sab in the order of rank. Any invalid tty values are skipped and those
not included in the tty_rank vector are subranked after. If
not provided, the function defaults to the rank given to the sab in the MRRANK
table in descending order.}
}
\value{
A table named by \code{sab} is written to the \code{crosswalk_schema}. If \code{sab}
value contained any punctuation, it is replaced with an underscore.
}
\description{
Subset the MRCONSO table by vocabulary and reduce the rows
to 1 \code{str} per \code{code}. All original fields from the MRCONSO
table are preserved. Crosswalks can then be performed by
\code{cui} between vocabularies. This function assumes that the core metathesaurus
tables are within a \code{umls} schema. Note that any source \code{sab}
value that contains punctuation other than an underscore
is replaced with an underscore to write tables with names that
Postgres can accept.
}
