% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_extension.R
\name{write_mrconso_sab_table}
\alias{write_mrconso_sab_table}
\title{Write a SAB Subset}
\usage{
write_mrconso_sab_table(
  sab,
  extension_schema,
  conn_fun = "pg13::local_connect(verbose = FALSE)",
  tty_rank,
  verbose = TRUE,
  render_sql = TRUE
)
}
\arguments{
\item{sab}{Vocabulary.}

\item{extension_schema}{Schema to which these extension tables are to be written.}

\item{tty_rank}{Vector of tty values for the given sab in the order of rank. If
not provided, the function defaults to the rank given to the sab in the MRRANK
table in descending order.}
}
\value{
A table named by \code{sab} is written to the \code{extension_schema}. If \code{sab}
value contained any punctuation, it is replaced with an underscore.
}
\description{
Filter the MRCONSO table for a given vocabulary and the STR
value designated as the main LABEL, which is an added field
in the table. This function assumes that the core metathesaurus
tables are within a \code{mth} schema.
}
\seealso{
\code{\link[stringr]{str_replace}}
\code{\link[SqlRender]{render}}
\code{\link[pg13]{send}},\code{\link[pg13]{query}}
}
