% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxnorm_map.R
\name{write_rxclass_table}
\alias{write_rxclass_table}
\title{Write RxClass Table}
\usage{
write_rxclass_table(
  conn,
  conn_fun = "pg13::local_connect()",
  schema = "mth",
  target_schema = "rxmap",
  rxclass_schema = "rxclass",
  mth_version,
  mth_release_dt,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  checks = ""
)
}
\description{
Take the RxClass classification table and map it to
the Ingredient maps generated in RxMap.
}
