% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxnorm_map.R
\name{write_rxnorm_all_maps}
\alias{write_rxnorm_all_maps}
\title{Write All RxNorm Maps}
\usage{
write_rxnorm_all_maps(
  conn,
  conn_fun = "pg13::local_connect()",
  schema = "mth",
  target_schema = "rxmap",
  mth_version,
  mth_release_dt,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  checks = ""
)
}
\description{
Each map is written to the given schema.
}
\concept{RxNorm Maps}
