% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxnorm_map.R
\name{write_rxnorm_ingredient_map}
\alias{write_rxnorm_ingredient_map}
\title{Write RxNorm Ingredient Map}
\usage{
write_rxnorm_ingredient_map(
  conn,
  conn_fun = "pg13::local_connect()",
  schema = "umls",
  target_schema = "rxmap",
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  checks = ""
)
}
\description{
Write RxNorm Ingredient Map
}
