% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxnorm_map.R
\name{write_rxnorm_map}
\alias{write_rxnorm_map}
\title{FUNCTION_TITLE}
\usage{
write_rxnorm_map(
  conn,
  conn_fun = "pg13::local_connect()",
  from_tty = c("BN", "BPCK", "DF", "DFG", "ET", "GPCK", "IN", "MIN", "PIN", "PSN",
    "SBD", "SBDC", "SBDF", "SBDG", "SCD", "SCDC", "SCDF", "SCDG", "SY", "TMSY"),
  to_tty = c("BN", "BPCK", "DF", "DFG", "ET", "GPCK", "IN", "MIN", "PIN", "PSN", "SBD",
    "SBDC", "SBDF", "SBDG", "SCD", "SCDC", "SCDF", "SCDG", "SY", "TMSY"),
  full_path = FALSE,
  schema = "mth",
  target_schema = "rxclass",
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  checks = ""
)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}

\item{conn_fun}{PARAM_DESCRIPTION, Default: 'pg13::local_connect()'}

\item{from_tty}{PARAM_DESCRIPTION, Default: c("BN", "BPCK", "DF", "DFG", "ET", "GPCK", "IN", "MIN", "PIN",
"PSN", "SBD", "SBDC", "SBDF", "SBDG", "SCD", "SCDC", "SCDF",
"SCDG", "SY", "TMSY")}

\item{to_tty}{PARAM_DESCRIPTION, Default: c("BN", "BPCK", "DF", "DFG", "ET", "GPCK", "IN", "MIN", "PIN",
"PSN", "SBD", "SBDC", "SBDF", "SBDG", "SCD", "SCDC", "SCDF",
"SCDG", "SY", "TMSY")}

\item{full_path}{PARAM_DESCRIPTION, Default: FALSE}

\item{schema}{PARAM_DESCRIPTION, Default: 'mth'}

\item{target_schema}{PARAM_DESCRIPTION, Default: 'rxclass'}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{render_sql}{PARAM_DESCRIPTION, Default: TRUE}

\item{render_only}{PARAM_DESCRIPTION, Default: FALSE}

\item{checks}{PARAM_DESCRIPTION, Default: ''}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[cli]{cli_alert}},\code{\link[cli]{cli_abort}}
\code{\link[rlang]{parse_expr}}
\code{\link[pg13]{dc}},\code{\link[pg13]{c("query", "query")}},\code{\link[pg13]{send}}
\code{\link[dplyr]{filter}}
\code{\link[glue]{glue}}
\code{\link[readr]{write_delim}}
}
\concept{RxNorm Class}
