% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxnorm_map.R
\name{write_rxnorm_path_lookup}
\alias{write_rxnorm_path_lookup}
\title{Write RxNorm Path Lookup}
\usage{
write_rxnorm_path_lookup(
  conn,
  conn_fun = "pg13::local_connect()",
  target_schema = "rxmap",
  table_name = "lookup_rxnorm_paths",
  umls_version,
  umls_release_dt,
  url = "https://lhncbc.nlm.nih.gov/RxNav/applications/RxNavViews.html#label:appendix",
  check_for_updates = FALSE,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  checks = ""
)
}
\arguments{
\item{conn}{Database connection via DatabaseConnector.}

\item{conn_fun}{Instead of supplying of a direct
database connection, an expression can be supplied that
automatically connects and disconnects the connection
once the function is run.}

\item{url}{Source URL. It is not hardcoded in case the URL changes in the future.
Default: 'https://lhncbc.nlm.nih.gov/RxNav/applications/RxNavViews.html#label:appendix'}

\item{verbose}{If TRUE, details on the activity are returned in the console, such as when the querying starting and was completed.}

\item{render_sql}{If TRUE, the SQL statement for the operation is returned in the console.}
}
\description{
Write the tibble returned by \code{read_rxnorm_paths()} to a
Postgres table.
}
\seealso{
Other RxNorm Map: 
\code{\link{get_rxnorm_ingredient_map}()},
\code{\link{get_rxnorm_map}()},
\code{\link{get_rxnorm_tty_lookup}()},
\code{\link{read_rxnorm_paths}()},
\code{\link{setup_rxmap}()},
\code{\link{write_rxnorm_map}()}
}
\concept{RxNorm Map}
