% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxnorm_map.R
\name{write_rxnorm_tty_lookup}
\alias{write_rxnorm_tty_lookup}
\title{Write RxNorm TTY Lookup}
\usage{
write_rxnorm_tty_lookup(
  conn,
  conn_fun = "pg13::local_connect()",
  schema = "mth",
  target_schema = "rxmap",
  mth_version,
  mth_release_dt,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  checks = ""
)
}
\description{
Write RxNorm TTY Lookup
}
