% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate_mysql_to_pg.R
\name{migrate_mysql_to_pg}
\alias{migrate_mysql_to_pg}
\title{FUNCTION_TITLE}
\usage{
migrate_mysql_to_pg(umls_conn, conn)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
https://evs.nci.nih.gov/evs-download/thesaurus-downloads
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pg13]{send}},\code{\link[pg13]{dropTable}}
\code{\link[SqlRender]{render}}
}
\concept{nci evs schema}
