% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metathesaurus-package.R
\docType{package}
\name{metathesaurus-package}
\alias{metathesaurus}
\alias{metathesaurus-package}
\title{metathesaurus: Setup UMLS Metathesaurus}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Setup UMLS Metathesaurus in MySQL5.5 or Postgres along with optional postprocessing steps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://meerapatelmd.github.io/metathesaurus/}
  \item \url{https://github.com/meerapatelmd/metathesaurus/}
  \item Report bugs at \url{https://github.com/meerapatelmd/metathesaurus/issues}
}

}
\author{
\strong{Maintainer}: Meera Patel \email{patelmeeray@gmail.com} (\href{https://orcid.org/0000-0002-3401-3948}{ORCID})

}
\keyword{internal}
