% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{rrfToCsv}
\alias{rrfToCsv}
\title{(Deprecated) RRF to CSV}
\usage{
rrfToCsv(path)
}
\description{
(Deprecated) RRF to CSV
}
\seealso{
\code{\link[progress]{progress_bar}}
\code{\link[cave]{strip_fn}}
\code{\link[readr]{read_delim}}
\code{\link[rubix]{rm_multibyte_chars}}
\code{\link[broca]{simply_write_csv}}
}
