% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mysql.R
\name{run_mysql}
\alias{run_mysql}
\title{Instantiate MySQL5.5}
\usage{
run_mysql(
  conn,
  mrconso_only = FALSE,
  omop_only = FALSE,
  english_only = TRUE,
  rrf_dir
)
}
\arguments{
\item{conn}{Database connection. For MySQL5.5, the connection made by RMySQL's dbConnect and for Postgres, the connection via DatabaseConnector.}

\item{mrconso_only}{Setup only the MRCONSO Table? Default: FALSE}

\item{omop_only}{Setup the MRCONSO, MRHIER, MRMAP, MRSMAP, MRSAT, AND MRREL Tables only? Default: FALSE}

\item{english_only}{Setup only the ENG Tables? PARAM_DESCRIPTION, Default: TRUE}

\item{rrf_dir}{Path to directory to the RRF Files either unpackaged by direct download or produced by Metamorphosys}
}
\description{
These functions setup the instance of the given DBMS. All SQL Statements that fail to execute are printed back in the console at the conclusion of execution. It is important to note that the schemas are dropped and rewritten with the run and all present tables will be lost.
}
\seealso{
\code{\link[preQL]{query}}
\code{\link[purrr]{map}},\code{\link[purrr]{map2}}
\code{\link[SqlRender]{render}}
\code{\link[stringr]{str_replace}},\code{\link[stringr]{str_remove}}
\code{\link[RMySQL]{character(0)}}
\code{\link[tibble]{as_tibble}}
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}},\code{\link[dplyr]{distinct}}
\code{\link[rubix]{filter_for}}
\code{\link[progress]{progress_bar}}

Other setup: 
\code{\link{run_pg}()}
}
\concept{setup}
