% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{runIndices}
\alias{runIndices}
\title{(Deprecated) Run Indices}
\usage{
runIndices(dbname = "umls", username, password, full = FALSE)
}
\arguments{
\item{path}{Path to the unpacked RRF files}
}
\description{
Run indices on the loaded data.
}
\keyword{internal}
