% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg.R
\name{setup_pg_mth}
\alias{setup_pg_mth}
\title{Instantiate Postgres (Deprecated)}
\usage{
setup_pg_mth(
  conn,
  conn_fun = "pg13::local_connect()",
  schema = "mth",
  path_to_rrfs,
  steps = c("reset_schema", "ddl_tables", "copy_rrfs", "add_indexes", "log",
    "setup_crosswalk"),
  mrconso_only = FALSE,
  omop_only = FALSE,
  english_only = TRUE,
  log_schema = "public",
  log_table_name = "setup_mth_log",
  log_version,
  log_release_date,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\description{
Instantiate Postgres (Deprecated)
}
\seealso{
\code{\link[pg13]{lsSchema}},\code{\link[pg13]{send}},\code{\link[pg13]{lsTables}},\code{\link[pg13]{dropTable}}
\code{\link[SqlRender]{render}}
\code{\link[tibble]{as_tibble}}
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}},\code{\link[dplyr]{distinct}}
\code{\link[stringr]{str_remove}}
\code{\link[progress]{progress_bar}}

Other setup: 
\code{\link{run_setup}()},
\code{\link{setup_mysql_mth}()}
}
\concept{setup}
