% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxnorm_map.R
\name{write_rxnorm_map}
\alias{write_rxnorm_map}
\title{Write RxNorm Map}
\usage{
write_rxnorm_map(
  conn,
  conn_fun = "pg13::local_connect()",
  from_tty = c("BN", "BPCK", "DF", "DFG", "ET", "GPCK", "IN", "MIN", "PIN", "PSN",
    "SBD", "SBDC", "SBDF", "SBDG", "SCD", "SCDC", "SCDF", "SCDG", "SY", "TMSY"),
  to_tty = c("BN", "BPCK", "DF", "DFG", "ET", "GPCK", "IN", "MIN", "PIN", "PSN", "SBD",
    "SBDC", "SBDF", "SBDG", "SCD", "SCDC", "SCDF", "SCDG", "SY", "TMSY"),
  full_path = FALSE,
  schema = "umls",
  target_schema = "rxmap",
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  checks = ""
)
}
\description{
Write RxNorm Map
}
\seealso{
Other RxNorm Map: 
\code{\link{get_rxnorm_ingredient_map}()},
\code{\link{get_rxnorm_map}()},
\code{\link{get_rxnorm_tty_lookup}()},
\code{\link{read_rxnorm_paths}()},
\code{\link{setup_rxmap}()},
\code{\link{write_rxnorm_path_lookup}()}
}
\concept{RxNorm Map}
